resource/script/pyproject.toml
[project]
name = "script"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "minio>=7.2.18",
    "python-dotenv>=1.1.1",
]

